Add hc-app command for building app bundles
