+++ crates/hc_bundle/src/bin/hc-app.rs
@@ -1,0 +1,12 @@
+use hc_bundle::cli::HcAppBundle;
+use structopt::StructOpt;
+
+/// Main `hc-app` executable entrypoint.
+#[tokio::main]
+pub async fn main() {
+    let opt = HcAppBundle::from_args();
+    if let Err(err) = opt.run().await {
+        eprintln!("hc-app: {}", err);
+        std::process::exit(1);
+    }
+}
